move EntryKind to entries
